refactor: move FormalDispatcherSequencialTester
